+++ packages/versions/src/lib/getBuiltinVersions.ts
@@ -4,6 +4,6 @@ export function getBuiltinVersions(): Versions {
   return {
     FUEL_CORE: '0.42.0',
     FORC: '0.67.0',
-    FUELS: '0.100.0',
+    FUELS: '0.100.1',
   };
 }