+++ packages/account/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+## 0.100.1
+
+### Patch Changes
+
+- fdefaba: fix: skip decoding external contract logs without JSON ABI
+- bb6a76c: chore: eliminate chain requests across concurrent provider instances
+  - @fuel-ts/abi-coder@0.100.1
+  - @fuel-ts/address@0.100.1
+  - @fuel-ts/crypto@0.100.1
+  - @fuel-ts/errors@0.100.1
+  - @fuel-ts/hasher@0.100.1
+  - @fuel-ts/math@0.100.1
+  - @fuel-ts/merkle@0.100.1
+  - @fuel-ts/transactions@0.100.1
+  - @fuel-ts/utils@0.100.1
+  - @fuel-ts/versions@0.100.1
+
 ## 0.100.0
 
 ### Minor Changes