+++ packages/create-fuels/CHANGELOG.md
@@ -1,5 +1,12 @@
 # create-fuels
 
+## 0.100.1
+
+### Patch Changes
+
+- @fuel-ts/errors@0.100.1
+- @fuel-ts/versions@0.100.1
+
 ## 0.100.0
 
 ### Patch Changes